Change path parameter for /v2/naics/group/ endpoint to use String instead of int.
The /v2/naics/group/ endpoint takes an Ask Kodiak NAICS group id path parameter to return a
NaicsGroup. The Ask Kodiak NAICS group id can be a NAICS code or a range of NAICS codes
(e.g. 31-33). To accept a range of NAICS codes, a string value specifying the range
can passed to the /v2/naics/group/ endpoint. In this SDK, this endpoint was initially
setup to only accept parameters of type int. So, the groupNumber parameter for the
getGroup method associated with the /v2/naics/group/ endpoint was changed from int
to String. An Ask Kodiak NAICS group id with a range can now be specified.