Do Not Encode "+" Symbol in Query Parameters
- The AskKodiak API requires a "+" symbol to delimit list items in some
query parameters. When "+" is encoded to "%2B", then the API does not
return results as expected.

- The current implementation of `getProductsEligibleForCode` uses the
`QueryMap` annotation and does not set the value of `encoded`. This
means that a default value of `false` is used, so Feign assumes that
the query parameters are not already encoded, and it encodes them,
including replacing "+" symbols with "%2B".

- For the sake of backward compatibility, add an overload of
 `getProductsEligibleForCode` to the`AskKodiak` interface that takes a
 `Map` instead of an `EligibleQuery` as the `QueryMap` object. Then
 made a default implementation of the existing
 `getProductsEligibleForCode` method that builds a `Map` from an
 `EligibleQuery` object (including encoding parameters except for "+"
 symbols) and calls the new overload method

 - Also upgraded the Feign dependency as there was an issue with
 `encoded = true` being respected on the `QueryMap` annotation in the
 old version